+++ libsecurity/EncryptedStorage.h
@@ -33,7 +33,7 @@ class EncryptedStorage : public Storage
 public:
     typedef std::shared_ptr<EncryptedStorage> Ptr;
     virtual ~EncryptedStorage(){};
-    Entries::Ptr select(h256 hash, int num, TableInfo::Ptr tableInfo, const std::string& key,
+    Entries::Ptr select(h256 hash, int64_t num, TableInfo::Ptr tableInfo, const std::string& key,
         Condition::Ptr condition = nullptr) override;
     size_t commit(h256 hash, int64_t num, const std::vector<TableData::Ptr>& datas) override;
